requirements.txt
Flask==2.2.2
pymongo==4.3.3
python-dotenv==0.21.1
black==22.12.0
Werkzeug==2.2.2